+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 0.17.0-0.8.0 (upcoming)
 
+* [PLT-3360] Bump cert-manager version to version 1.19.1 and cluster-api version to version 1.10.8
 * [PLT-3362] bump azurefile-csi version to 1.34.1
 * [PLT-3363] bump azuredisk-csi version to 1.33.5
 * [PLT-3427][Fix] Condición de Carrera en cloud-provisioner